+++ src/everos/memory/strategies/__init__.py
@@ -5,6 +5,7 @@ External usage:
         extract_agent_case,
         extract_agent_skill,
         extract_atomic_facts,
+        extract_decision,
         extract_foresight,
         extract_user_profile,
         reflect_episodes,
@@ -16,6 +17,7 @@ External usage:
 from .extract_agent_case import extract_agent_case as extract_agent_case
 from .extract_agent_skill import extract_agent_skill as extract_agent_skill
 from .extract_atomic_facts import extract_atomic_facts as extract_atomic_facts
+from .extract_decision import extract_decision as extract_decision
 from .extract_foresight import extract_foresight as extract_foresight
 from .extract_user_profile import extract_user_profile as extract_user_profile
 from .reflect_episodes import reflect_episodes as reflect_episodes
@@ -30,6 +32,7 @@ __all__ = [
     "extract_agent_case",
     "extract_agent_skill",
     "extract_atomic_facts",
+    "extract_decision",
     "extract_foresight",
     "extract_user_profile",
     "reflect_episodes",